Updating MessageFormat, no longer requiring external locale files
The 0.2.x branch of MessageFormat got its number formatting rules from custom rule files that had to be included in the browser.
The 0.3.x branch now uses information build into the browser (or so I've gathered from the limited info in their docs), and doesn't provide the locale files anymore.
Since we no longer need to load external files, this library can be a lot simpler, both internally and using it in an FRA.
Nothing using it needs to maintain a MessageFormat dependency anymore.